+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * interrupt support with gpioSetISRFunc
+  * throw errors when pigpio returns errors
 
 0.0.0 / Oct 14 2015
 ===================